refactoring e2e testsuite code to group together logically
